requirements.txt
arrow==0.7.0
beautifulsoup4==4.4.1
disqus==0.0.4
Django==1.8.9
django-disqus==0.5
django-emoticons==1.1.2
django-forms-bootstrap==3.0.1
django-material==0.7.0
django-paypal==0.2.7
django-tinymce==2.2.0
djangorestframework==3.3.3
funcsigs==1.0.0
material==0.1
mock==2.0.0
ordereddict==1.1
paypal==1.2.5
pbr==1.9.1
Pillow==3.1.1
python-dateutil==2.4.2
pytz==2016.3
requests==2.9.1
reusable-blog-app==1.0.0
six==1.10.0
sqlparse==0.1.19
stripe==1.32.0
wheel==0.24.0
